// src/types.ts
export type NavBarProps = {
  links: string[];
  sign_in: string;
};

export type HeroProps = {
  title_1: string;
  title_2: string;
  description_1: string;
  description_2: string;
  description_3: string;
  button_txt: string;
};

export type GenerateCTAProps = {
  title: string;
  description_1: string;
  description_2: string;
  button_txt: string;
};

export type AssistantDocsProps = {
  top_title: string;
  title: string;
  description: string;
  button_txt: string;
};

export type InfoBlocksProps = {
  top_title_1: string;
  title_1: string;
  description_1: string;
  top_title_2: string;
  title_2: string;
  description_2: string;
};

export type FooterProps = {
  company_menu_title: string;
  product_menu_title: string;
  company_menu: string[];
  product_menu: string[];
  established: string;
  headline_1: string;
  headline_2: string;
  languages: string[];
};